requirements.txt
flask
mysql-connector-python
authlib
requests